update dap_reset_target_hw() with state
move DAP_SAM and DAP_SAMx5 to separated header
