+++ Adafruit_DAP_SAMx5.cpp
@@ -106,9 +106,9 @@ bool Adafruit_DAP_SAMx5::select(uint32_t *found_id) {
   uint32_t DAP_DSU_did;
 
   // Stopping the core fails on locked SAM D21/51, when not doing an Extended reset first.
-  // As per the ataradov/edbg source code, for SAMD MCUs which is locked by having the Security Bit set, one must enter Reset Extension mode
-  // before issuing reads through DAP. Write mostly requires the Security Bit to be cleared by executing a Chip Erase, before entering Reset Extension mode
-  // once more, followed by a SWD reconnect.
+  // As per the ataradov/edbg source code, for SAMD MCUs which is locked by having the Security Bit set,
+  // one must enter Reset Extension mode before issuing reads through DAP. Write mostly requires the Security Bit
+  // to be cleared by executing a Chip Erase, before entering Reset Extension mode once more, followed by a SWD reconnect.
 
   resetWithExtension();
